Match naming of other files and style code
